Fixes for left navigation bar
This change has 3 fixes for the sidebar navigation. The first fix allows for a different navigation title than the page title. The second fix allows the same for collection objects in the navigation. The third fix introduces a break in the for loop after the first notification item is read.