Update 01_common_policy_rekey.md
@ryancdickson An **attempt** at some simpler language.  I didn't update the last date. 

We are using jargon.  Although the jargon is technically correct, it can be confusing.  IMO, its okay to keep the jargon debates for the water cooler - and instead use strong, simple terms that will achieve the goal.  The goal is: You need to do X (inform).  Here is how to do X (actions).